requirements.txt
# Essential packages
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
nltk>=3.8.1
textblob>=0.17.1
Flask>=3.0.0
Flask-CORS>=4.0.0
vaderSentiment>=3.3.2
PyJWT>=2.8.0
python-dotenv>=1.0.0

# Optional: For AI-powered summaries (app works without these using rule-based fallback)
# transformers>=4.30.0
# torch>=2.0.0

